Move huggingface_api_manager.py to util folder
